move GTClientPollutionMap class to pollution package
