+++ src/main/java/gregtech/common/pollution/PollutionRenderer.java
@@ -23,7 +23,6 @@ import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 import gregtech.GTMod;
 import gregtech.common.entities.EntityFXPollution;
-import gregtech.common.misc.GTClientPollutionMap;
 
 @SideOnly(Side.CLIENT)
 public class PollutionRenderer {
